PowerControl: Move `Menu.cs` into `Menu/`
